+++ packages/analytics/src/components/Map/components/MapLayer/components/ThematicLayer/hooks/config.ts
@@ -1,7 +1,7 @@
 import { find } from "lodash";
 import { useContext } from "react";
-import { MapLayersContext } from "../../../../../state/index.js";
-import { CustomThematicLayer } from "../../../interfaces/index.js";
+import { MapLayersContext } from "../../../../../state";
+import { CustomThematicLayer } from "../../../interfaces";
 
 export default function useThematicLayer(
 	layerId: string,