+++ packages/scorecard/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hisptz/dhis2-scorecard
 
+## 1.2.39
+
+### Patch Changes
+
+- Updated dependencies
+    - @hisptz/dhis2-ui@2.0.54
+    - @hisptz/dhis2-analytics@2.2.2
+
 ## 1.2.38
 
 ### Patch Changes